Внес изменения

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -96,6 +96,9 @@ public class UserService {
 
     public void deleteUser(long userId) {
         checkUserExists(userId);
+        if (userStorage.findById(userId) == null) {
+            throw new NotFoundException("Пользователь с ID [" + userId + "] не найден");
+        }
         friendshipStorage.removeAllByUserId(userId);
         userStorage.deleteUser(userId);
